PageType Offset 0, DebugPage().

+++ Dbarone.Net.Database/Core/Pages/PageType.cs
@@ -1,7 +1,7 @@
 /// <summary>
 /// The page type.
 /// </summary>
-public enum PageType : int
+public enum PageType : Byte
 {
     /// <summary>
     /// A page type that stores raw leaf level data.
